requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
pytz==2024.1
django-extensions==3.2.3